added: requests controller
