+++ src/controllers/selectedAccount/selectedAccount.ts
@@ -27,8 +27,6 @@ import { calculateSelectedAccountPortfolio } from '../../libs/selectedAccount/se
 // eslint-disable-next-line import/no-cycle
 import { AccountsController } from '../accounts/accounts'
 // eslint-disable-next-line import/no-cycle
-import { ActionsController } from '../actions/actions'
-// eslint-disable-next-line import/no-cycle
 import { DefiPositionsController } from '../defiPositions/defiPositions'
 import EventEmitter from '../eventEmitter/eventEmitter'
 import { NetworksController } from '../networks/networks'
@@ -57,8 +55,6 @@ export class SelectedAccountController extends EventEmitter {
   #portfolio: PortfolioController | null = null
 
   #defiPositions: DefiPositionsController | null = null
-
-  #actions: ActionsController | null = null
 
   #networks: NetworksController | null = null
 
@@ -147,19 +143,16 @@ export class SelectedAccountController extends EventEmitter {
   initControllers({
     portfolio,
     defiPositions,
-    actions,
     networks,
     providers
   }: {
     portfolio: PortfolioController
     defiPositions: DefiPositionsController
-    actions: ActionsController
     networks: NetworksController
     providers: ProvidersController
   }) {
     this.#portfolio = portfolio
     this.#defiPositions = defiPositions
-    this.#actions = actions
     this.#networks = networks
     this.#providers = providers
 
@@ -291,10 +284,6 @@ export class SelectedAccountController extends EventEmitter {
       this.#portfolio.getPendingPortfolioState(this.account.addr)
     )
 
-    const hasSignAccountOp = !!this.#actions?.visibleActionsQueue.filter(
-      (action) => action.type === 'accountOp'
-    )
-
     const {
       selectedAccountPortfolio: newSelectedAccountPortfolio,
       selectedAccountPortfolioByNetworks: newSelectedAccountPortfolioByNetworks
@@ -304,7 +293,6 @@ export class SelectedAccountController extends EventEmitter {
       structuredClone(this.#portfolioByNetworks),
       this.portfolioStartedLoadingAtTimestamp,
       defiPositionsAccountState,
-      hasSignAccountOp,
       this.#isPortfolioLoadingFromScratch
     )
